requirements.txt
certifi==2018.4.16
dataclasses==0.6
dj-database-url==0.5.0
Django==2.0.7
django-filter==1.1.0
django-heroku==0.3.1
djangorestframework==3.8.2
djongo==1.2.29
Markdown==2.6.11
mongoengine==0.15.0
psycopg2==2.7.5
pymongo==3.7.0
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
whitenoise==3.3.1
